requirements2.txt
absl-py==2.1.0
accelerate==0.30.1
agentlab==0.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiohttp-cors==0.7.0
aiolimiter==1.2.1
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
av==14.2.0
babel==2.17.0
beartype==0.12.0
beautifulsoup4==4.13.3
bitsandbytes==0.43.1
black==25.1.0
blacken-docs==1.19.1
blake3==1.0.4
bleach==6.2.0
blinker==1.9.0
browsergym==0.13.3
browsergym-assistantbench==0.13.3
browsergym-core==0.13.3
browsergym-experiments==0.13.3
browsergym-miniwob==0.13.3
browsergym-visualwebarena==0.13.3
browsergym-webarena==0.13.3
browsergym-workarena==0.4.1
cachetools==5.5.2
certifi==2022.12.7
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==2.1.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorful==0.5.6
comm==0.2.2
compressed-tensors==0.9.1
contexttimer==0.3.3
contourpy==1.3.1
cupy-cuda12x==13.4.0
cycler==0.12.1
dask==2025.2.0
dataclasses-json==0.6.7
datasets==3.3.2
debugpy==1.8.12
decorator==5.2.1
deepspeed==0.13.1
defusedxml==0.7.1
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distributed==2025.2.0
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
einops==0.8.1
email_validator==2.2.0
english-words==2.0.1
evaluate==0.4.3
execnet==2.1.1
executing==2.2.0
Faker==36.1.1
Farama-Notifications==0.0.4
fastapi==0.115.11
fastapi-cli==0.0.7
fastjsonschema==2.21.1
fastrlock==0.8.3
ffmpy==0.5.0
filelock==3.17.0
flaky==3.8.1
Flask==3.1.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.6.1
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.1
gradio==5.20.0
gradio_client==1.7.2
greenlet==3.1.1
groovy==0.1.2
grpcio==1.70.0
gymnasium==1.1.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
identify==2.6.9
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
iniconfig==2.0.0
interegular==0.3.3
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.5.1
jaxlib==0.5.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.43
langchain-text-splitters==0.3.6
langsmith==0.3.13
lark==1.2.2
lazy_loader==0.4
Levenshtein==0.27.1
libvisualwebarena==0.0.15
libwebarena==0.0.4
linkify-it-py==2.0.3
llvmlite==0.43.0
lm-format-enforcer==0.10.11
locket==1.0.0
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.16.0
mistral_common==1.5.3
mistune==3.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.3
nltk==3.9.1
nodeenv==1.9.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.4.2
openai==1.65.2
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
orjson==3.10.15
outcome==1.3.0.post0
outlines==0.1.11
outlines_core==0.1.26
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
partial-json-parser==0.2.1.1.post5
pathspec==0.12.1
peft==0.14.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
playwright==1.50.0
pluggy==1.5.0
portalocker==3.1.1
pre_commit==4.1.0
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycocotools==2.0.8
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
pyee==12.1.1
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.1
PySocks==1.7.1
pytest==7.3.2
pytest-base-url==2.1.0
pytest-playwright==0.7.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
qwen-vl-utils==0.0.10
RapidFuzz==3.12.2
ray==2.40.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.23.1
rsa==4.9
ruff==0.9.9
sacrebleu==2.5.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.12.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spaces==0.32.0
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.46.0
sympy==1.13.1
tabulate==0.9.0
tblib==3.0.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
terminado==0.18.1
text-generation==0.7.0
text-unidecode==1.3
textual==2.1.2
threadpoolctl==3.5.0
tifffile==2025.2.18
tiktoken==0.9.0
tinycss2==1.4.0
tokenize_rt==6.1.0
tokenizers==0.21.0
tomlkit==0.13.2
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
trio==0.29.0
trio-websocket==0.12.2
triton==3.1.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20250301
types-tqdm==4.67.0.20250301
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
ultralytics==8.3.82
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.3
vllm==0.7.3
wandb==0.19.7
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
weblinx==0.3.2
weblinx-browsergym==0.0.1.dev14
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xformers==0.0.28.post3
xgrammar==0.1.11
xxhash==3.5.0
yarl==1.18.3
zict==3.0.0
zipp==3.21.0
zstandard==0.23.0
